Move data-stuff to data-commands
