+++ freqtrade/commands/__init__.py
@@ -1,10 +1,10 @@
-from .hyperopt_commands import (start_hyperopt_list, start_hyperopt_show)  # noqa: 401
-
-from .list_commands import (start_list_exchanges,  # noqa: F401
+from freqtrade.commands.data_commands import start_download_data  # noqa: F401
+from freqtrade.commands.deploy_commands import (start_create_userdir,  # noqa: F401
+                              start_new_hyperopt, start_new_strategy)
+from freqtrade.commands.hyperopt_commands import (start_hyperopt_list,  # noqa: F401
+                                start_hyperopt_show)
+from freqtrade.commands.list_commands import (start_list_exchanges,  # noqa: F401
                             start_list_markets, start_list_strategies,
                             start_list_timeframes)
-from .utils import setup_utils_configuration  # noqa: F401
-from .utils import (start_download_data,  # noqa: F401
-                    start_test_pairlist)
-from .deploy_commands import (start_new_hyperopt, start_new_strategy, start_create_userdir)  # noqa: F401
-from .trade_commands import start_trading  # noqa: F401
+from freqtrade.commands.trade_commands import start_trading  # noqa: F401
+from freqtrade.commands.utils import setup_utils_configuration, start_test_pairlist  # noqa: F401
